refactor: isolate v-inject into dedicated module
Extract v-inject state machine from assets/hooks.ts into assets/inject.ts
and shared DOM helpers into assets/attrs.ts. Core hook drops from 378 to
~115 lines; v-inject logic no longer clutters the general hook path for
users who don't adopt Vue-rendered layouts.

On the server side, collapse InjectedSSR's two entry points into a single
prepare/3 and compute inject_config once per vue/1 call instead of four
times per render.
